Move tests in AbstractTestRest to nested classes
Just moving tests, not changing tests - removing the "monstrosity" of that 2500 line class.
