+++ servers/jax-rs-tests/src/main/java/org/projectnessie/jaxrs/AbstractRestInvalidRefs.java
@@ -1,0 +1,163 @@
+/*
+ * Copyright (C) 2022 Dremio
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.projectnessie.jaxrs;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.projectnessie.model.Validation.REF_NAME_MESSAGE;
+
+import com.google.common.collect.ImmutableList;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.function.Supplier;
+import java.util.stream.IntStream;
+import org.junit.jupiter.api.Test;
+import org.projectnessie.client.api.NessieApiV1;
+import org.projectnessie.client.rest.NessieBadRequestException;
+import org.projectnessie.error.BaseNessieClientServerException;
+import org.projectnessie.error.NessieNotFoundException;
+import org.projectnessie.model.Branch;
+import org.projectnessie.model.Content;
+import org.projectnessie.model.ContentKey;
+import org.projectnessie.model.EntriesResponse;
+import org.projectnessie.model.LogResponse;
+
+public abstract class AbstractRestInvalidRefs extends AbstractRest {
+  protected AbstractRestInvalidRefs(Supplier<NessieApiV1> api) {
+    super(api);
+  }
+
+  @Test
+  public void testInvalidNamedRefs() {
+    ContentKey key = ContentKey.of("x");
+    String invalidRef = "1234567890123456";
+
+    assertThatThrownBy(() -> getApi().getCommitLog().refName(invalidRef).get())
+        .isInstanceOf(NessieBadRequestException.class)
+        .hasMessageContaining("Bad Request (HTTP/400):")
+        .hasMessageContaining(REF_NAME_MESSAGE);
+
+    assertThatThrownBy(() -> getApi().getEntries().refName(invalidRef).get())
+        .isInstanceOf(NessieBadRequestException.class)
+        .hasMessageContaining("Bad Request (HTTP/400):")
+        .hasMessageContaining(REF_NAME_MESSAGE);
+
+    assertThatThrownBy(() -> getApi().getContent().key(key).refName(invalidRef).get())
+        .isInstanceOf(NessieBadRequestException.class)
+        .hasMessageContaining("Bad Request (HTTP/400):")
+        .hasMessageContaining(REF_NAME_MESSAGE);
+
+    assertThatThrownBy(() -> getApi().getContent().refName(invalidRef).key(key).get())
+        .isInstanceOf(NessieBadRequestException.class)
+        .hasMessageContaining("Bad Request (HTTP/400):")
+        .hasMessageContaining(REF_NAME_MESSAGE);
+  }
+
+  @Test
+  public void testValidHashesOnValidNamedRefs() throws BaseNessieClientServerException {
+    Branch branch = createBranch("testValidHashesOnValidNamedRefs");
+
+    int commits = 10;
+
+    String currentHash = branch.getHash();
+    createCommits(branch, 1, commits, currentHash);
+    LogResponse entireLog = getApi().getCommitLog().refName(branch.getName()).get();
+    assertThat(entireLog).isNotNull();
+    assertThat(entireLog.getLogEntries()).hasSize(commits);
+
+    EntriesResponse allEntries = getApi().getEntries().refName(branch.getName()).get();
+    assertThat(allEntries).isNotNull();
+    assertThat(allEntries.getEntries()).hasSize(commits);
+
+    List<ContentKey> keys = new ArrayList<>();
+    IntStream.range(0, commits).forEach(i -> keys.add(ContentKey.of("table" + i)));
+
+    // TODO: check where hashOnRef is set
+    Map<ContentKey, Content> allContent =
+        getApi().getContent().keys(keys).refName(branch.getName()).get();
+
+    for (int i = 0; i < commits; i++) {
+      String hash = entireLog.getLogEntries().get(i).getCommitMeta().getHash();
+      LogResponse log = getApi().getCommitLog().refName(branch.getName()).hashOnRef(hash).get();
+      assertThat(log).isNotNull();
+      assertThat(log.getLogEntries()).hasSize(commits - i);
+      assertThat(ImmutableList.copyOf(entireLog.getLogEntries()).subList(i, commits))
+          .containsExactlyElementsOf(log.getLogEntries());
+
+      EntriesResponse entries =
+          getApi().getEntries().refName(branch.getName()).hashOnRef(hash).get();
+      assertThat(entries).isNotNull();
+      assertThat(entries.getEntries()).hasSize(commits - i);
+
+      int idx = commits - 1 - i;
+      ContentKey key = ContentKey.of("table" + idx);
+      Content c =
+          getApi().getContent().key(key).refName(branch.getName()).hashOnRef(hash).get().get(key);
+      assertThat(c).isNotNull().isEqualTo(allContent.get(key));
+    }
+  }
+
+  @Test
+  public void testUnknownHashesOnValidNamedRefs() throws BaseNessieClientServerException {
+    Branch branch = createBranch("testUnknownHashesOnValidNamedRefs");
+    String invalidHash = "1234567890123456";
+
+    int commits = 10;
+
+    String currentHash = branch.getHash();
+    createCommits(branch, 1, commits, currentHash);
+    assertThatThrownBy(
+            () -> getApi().getCommitLog().refName(branch.getName()).hashOnRef(invalidHash).get())
+        .isInstanceOf(NessieNotFoundException.class)
+        .hasMessageContaining(
+            String.format(
+                "Could not find commit '%s' in reference '%s'.", invalidHash, branch.getName()));
+
+    assertThatThrownBy(
+            () -> getApi().getEntries().refName(branch.getName()).hashOnRef(invalidHash).get())
+        .isInstanceOf(NessieNotFoundException.class)
+        .hasMessageContaining(
+            String.format(
+                "Could not find commit '%s' in reference '%s'.", invalidHash, branch.getName()));
+
+    assertThatThrownBy(
+            () ->
+                getApi()
+                    .getContent()
+                    .key(ContentKey.of("table0"))
+                    .refName(branch.getName())
+                    .hashOnRef(invalidHash)
+                    .get())
+        .isInstanceOf(NessieNotFoundException.class)
+        .hasMessageContaining(
+            String.format(
+                "Could not find commit '%s' in reference '%s'.", invalidHash, branch.getName()));
+
+    assertThatThrownBy(
+            () ->
+                getApi()
+                    .getContent()
+                    .key(ContentKey.of("table0"))
+                    .refName(branch.getName())
+                    .hashOnRef(invalidHash)
+                    .get())
+        .isInstanceOf(NessieNotFoundException.class)
+        .hasMessageContaining(
+            String.format(
+                "Could not find commit '%s' in reference '%s'.", invalidHash, branch.getName()));
+  }
+}
